refactor: 프로젝트 구조에 맞게 테스트 분리
