+++ src/main/java/racingcar/controller/RaceController.java
@@ -49,5 +49,4 @@ public class RaceController {
         outputView.printWinners(cars);
     }
 
-
 }
