+++ src/test/java/racingcar/controller/RaceControllerTest.java
@@ -1,0 +1,16 @@
+package racingcar.controller;
+
+import camp.nextstep.edu.missionutils.test.NsTest;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
+import static org.assertj.core.api.Assertions.assertThat;
+
+class RaceControllerTest extends NsTest {
+
+    @Override
+    protected void runMain() {
+
+    }
+}
