+++ src/test/java/racingcar/view/IoTest.java
@@ -1,14 +1,9 @@
-package racingcar;
+package racingcar.view;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.controller.CarController;
-import racingcar.controller.RaceController;
-import racingcar.controller.MoveController;
 import racingcar.model.Car;
-import racingcar.view.InputView;
-import racingcar.view.OutputView;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,11 +11,10 @@ import java.util.List;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class RaceControllerTest extends NsTest {
+@DisplayName("입출력 테스트")
+class IoTest extends NsTest {
     private InputView inputView = new InputView();
     private OutputView outputView = new OutputView();
-    private CarController carController = new CarController();
-    private MoveController moveController = new MoveController();
 
     @Test
     @DisplayName("자동차 이름을 입력받는다")
@@ -32,43 +26,12 @@ class RaceControllerTest extends NsTest {
     }
 
     @Test
-    @DisplayName("쉼표를 기준으로 자동차 이름을 분리한다")
-    void splitCarNames() {
-        assertThat(carController.splitCarNames("pobi,woni,jun")).isEqualTo(List.of("pobi", "woni", "jun"));
-    }
-
-    @Test
-    @DisplayName("단일 자동차 객체를 생성한다")
-    void createSingleCar() {
-        assertThat(carController.createSingleCar("pobi")).isInstanceOf(Car.class);
-    }
-
-    @Test
-    @DisplayName("모든 자동차 객체를 생성한다")
-    void createAllCars() {
-        assertThat(carController.createAllCars(List.of("pobi", "woni", "jun")).size()).isEqualTo(3);
-    }
-
-    @Test
     @DisplayName("시도할 횟수를 입력받는다")
     void playCountInput() {
         assertSimpleTest(() -> {
             run("5");
             assertThat(inputView.receivePlayCount()).isEqualTo("5");
         });
-    }
-
-    @Test
-    @DisplayName("생성된 랜덤 숫자가 범위 내인지 확인한다")
-    void checkRandomNumberRange() {
-        assertThat(moveController.createRandomNumber()).isBetween(0, 9);
-    }
-
-    @Test
-    @DisplayName("랜덤 숫자가 4이상이면 전진한다")
-    void isMove() {
-        assertThat(moveController.isMove(4)).isTrue();
-        assertThat(moveController.isMove(3)).isFalse();
     }
 
     @Test
